requirements.txt
PyYAML
requests
Pillow
matplotlib
